监听函数重构/监听准备器
