+++ boots/simboot-api/src/main/kotlin/love/forte/simboot/interceptor/AnnotatedEventListenerInterceptor.kt
@@ -19,6 +19,7 @@ package love.forte.simboot.interceptor
 
 import love.forte.simbot.Api4J
 import love.forte.simbot.event.BlockingEventListenerInterceptor
+import love.forte.simbot.event.EventListener
 import love.forte.simbot.event.EventListenerInterceptor
 import love.forte.simbot.event.EventResult
 
@@ -31,6 +32,10 @@ import love.forte.simbot.event.EventResult
  * 而是通过 [@Interceptor(...)][love.forte.simboot.annotation.Interceptor]
  * 配合 [@Listener][love.forte.simboot.annotation.Listener] 进行使用。
  *
+ *
+ * [AnnotatedEventListenerInterceptor] 相对于 [EventListenerInterceptor] 比较特殊，它是针对并作用于某个**具体**的监听函数的，
+ * 因此它无法在比监听函数更上层的交流来处理它，进而决定了 [AnnotatedEventListenerInterceptor] 将只支持对 [EventListener.invoke] 进行拦截，
+ * 也就是说 [AnnotatedEventListenerInterceptor] 只会在 [EventListener.match] 匹配成功后才会执行。
  *
  * @see love.forte.simboot.annotation.Interceptor
  * @see EventListenerInterceptor
