+++ boots/simboot-api/src/main/kotlin/love/forte/simboot/listener/GenericListener.kt
@@ -18,9 +18,9 @@ package love.forte.simboot.listener
 
 import love.forte.simbot.ID
 import love.forte.simbot.event.Event
+import love.forte.simbot.event.EventListener
 import love.forte.simbot.event.EventListenerProcessingContext
 import love.forte.simbot.event.EventResult
-import love.forte.simbot.event.MatchableEventListener
 import love.forte.simbot.utils.runWithInterruptible
 
 /**
@@ -34,11 +34,11 @@ import love.forte.simbot.utils.runWithInterruptible
  *
  * @author ForteScarlet
  */
-public interface GenericBootEventListener : MatchableEventListener {
+public interface GenericBootEventListener : EventListener {
     override val id: ID
     override val isAsync: Boolean
     override fun isTarget(eventType: Event.Key<*>): Boolean
     override suspend fun match(context: EventListenerProcessingContext): Boolean
-    override suspend fun directInvoke(context: EventListenerProcessingContext): EventResult
+    override suspend fun invoke(context: EventListenerProcessingContext): EventResult
 }
