+++ boots/simboot-core-annotation/src/main/kotlin/love/forte/simboot/annotation/StandardTextContentProcessorInterceptors.kt
@@ -18,7 +18,6 @@
 package love.forte.simboot.annotation
 
 import love.forte.annotationtool.AnnotationMapper
-import love.forte.simboot.listener.EventListenerTextContentProcessor
 import love.forte.simboot.listener.StandardTextContentProcessor
 import love.forte.simbot.ExperimentalSimbotApi
 import love.forte.simbot.PriorityConstant
@@ -28,22 +27,21 @@ import love.forte.simbot.event.EventListenerProcessingContext
 /**
  * 对 [EventListenerProcessingContext.textContent] 进行前置处理, 清除前后空格（如果不为null的话）。
  *
- * [ContentTrim] 是 [Interceptor] 的变体，且在默认情况下，[ContentTrim] 的优先级是 **最高** 的。
+ * [ContentTrim] 是 [Preparator] 的变体，且在默认情况下，[ContentTrim] 的优先级是 **最高** 的。
  *
  * 不建议 [ContentTrim] 与 [ContentToNull] 一起使用。
  *
  * @see Interceptor
- * @see EventListenerTextContentProcessor
  * @see StandardTextContentProcessor.Trim
  */
 @ExperimentalSimbotApi
-@Interceptor(type = StandardTextContentProcessor.Trim::class, priority = PriorityConstant.FIRST)
+@Preparator(StandardTextContentProcessor.Trim::class, priority = PriorityConstant.FIRST)
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.ANNOTATION_CLASS)
 @MustBeDocumented
-@AnnotationMapper(Interceptor::class)
+@AnnotationMapper(Preparator::class)
 public annotation class ContentTrim(
-    @get:AnnotationMapper.Property(value = "priority", target = Interceptor::class)
-    val priority: Int = PriorityConstant.FIRST
+    @get:AnnotationMapper.Property(value = "priority", target = Preparator::class)
+    val priority: Int = PriorityConstant.FIRST,
 )
 
 /**
@@ -53,17 +51,16 @@ public annotation class ContentTrim(
  * 不建议 [ContentTrim] 与 [ContentToNull] 一起使用, 或者说，不建议 [ContentToNull] 与其他任何操作 [EventListenerProcessingContext.textContent] 的相关内容放在一起使用。
  *
  * @see Interceptor
- * @see EventListenerTextContentProcessor
  * @see StandardTextContentProcessor.Null
  */
 @ExperimentalSimbotApi
-@Interceptor(type = StandardTextContentProcessor.Null::class, priority = PriorityConstant.FIRST)
+@Preparator(StandardTextContentProcessor.Null::class, priority = PriorityConstant.FIRST)
 @Target(AnnotationTarget.FUNCTION, AnnotationTarget.ANNOTATION_CLASS)
 @MustBeDocumented
-@AnnotationMapper(Interceptor::class)
+@AnnotationMapper(Preparator::class)
 public annotation class ContentToNull(
-    @get:AnnotationMapper.Property(value = "priority", target = Interceptor::class)
-    val priority: Int = PriorityConstant.FIRST
+    @get:AnnotationMapper.Property(value = "priority", target = Preparator::class)
+    val priority: Int = PriorityConstant.FIRST,
 )
