+++ boots/simboot-core/src/main/kotlin/love/forte/simboot/core/listener/FunctionalListenerProcessContext.kt
@@ -18,7 +18,6 @@ package love.forte.simboot.core.listener
 
 import love.forte.di.BeanContainer
 import love.forte.simboot.core.binder.BinderManager
-import love.forte.simboot.core.interceptor.FunctionalListenerInterceptor
 import kotlin.reflect.KFunction
 
 /**
@@ -56,10 +55,5 @@ public data class FunctionalListenerProcessContext(
      */
     val beanContainer: BeanContainer,
     
-    // TODO
-    /**
-     * 监听函数内部的call前拦截。
-     */
-    val functionCallInterceptors: List<FunctionalListenerInterceptor> = emptyList(), // TODO
 )
