+++ cores/simbot-core/src/main/kotlin/love/forte/simbot/core/event/CoreIntercepts.kt
@@ -19,79 +19,5 @@
 
 package love.forte.simbot.core.event
 
-import love.forte.simbot.event.EventListener
-import love.forte.simbot.event.EventListenerInterceptor
-import love.forte.simbot.event.MatchableEventListener
 
-
-// TODO update doc
-
-/**
- *
- *
- */
-public operator fun EventListener.plus(interceptors: Collection<EventListenerInterceptor>): EventListener {
-    if (interceptors.isEmpty()) return this
-    if (this !is MatchableEventListener) {
-        val entrance = EventInterceptEntrance.eventListenerInterceptEntrance(interceptors)
-        return proxy { listener ->
-            entrance.doIntercept(this) { context0 ->
-                listener(context0)
-            }
-        }
-    }
-    
-    TODO()
-    
-   return plus(interceptors)
-}
-
-// TODO update doc
-
-/**
- *
- *
- */
-public operator fun MatchableEventListener.plus(interceptors: Collection<EventListenerInterceptor>): MatchableEventListener {
-    if (interceptors.isEmpty()) return this
-    
-    val groupByIsAfterMatch = interceptors.groupBy { it.point == EventListenerInterceptor.Point.AFTER_MATCH }
-    
-    val normalInterceptors = groupByIsAfterMatch[false] ?: emptyList()
-    val afterMatchInterceptors = groupByIsAfterMatch[true] ?: emptyList()
-    
-    
-    val normalEntrance = EventInterceptEntrance.eventListenerInterceptEntrance(normalInterceptors)
-    val afterMatchEntrance = EventInterceptEntrance.eventListenerInterceptEntrance(afterMatchInterceptors)
-    
-    return proxy({ listener ->
-        TODO()
-    }) { listener ->
-        TODO()
-    }
-    
-    // return proxy { listener ->
-    //     normalEntrance.doIntercept(this) { context0 ->
-    //         if (listener.match(context0)) {
-    //             afterMatchEntrance.doIntercept(context0) { context1 ->
-    //                 listener(context1)
-    //             }
-    //         } else {
-    //             EventResult.invalid()
-    //         }
-    //     }
-    // }
-}
-
-
-// internal class EventListenerWithInterceptor(
-//     private val listener: EventListener,
-//     interceptors: Collection<EventListenerInterceptor>
-// ) : EventListener by listener {
-//     private val entrance = EventListenerIteratorInterceptEntrance(listener, interceptors)
-//     override suspend fun invoke(context: EventListenerProcessingContext): EventResult {
-//         return entrance.doIntercept(context)
-//     }
-//
-// }
