+++ cores/simbot-core/src/main/kotlin/love/forte/simbot/core/event/SimpleListenerProxy.kt
@@ -19,11 +19,14 @@
 package love.forte.simbot.core.event
 
 import love.forte.simbot.ID
-import love.forte.simbot.event.*
+import love.forte.simbot.event.Event
+import love.forte.simbot.event.EventListener
+import love.forte.simbot.event.EventListenerProcessingContext
+import love.forte.simbot.event.EventResult
 import org.slf4j.Logger
 
 /**
- * 提供一个匹配函数，将当前监听函数转化为一个 [MatchableEventListener].
+ * 提供一个匹配函数，将当前监听函数根据新的匹配函数转化为一个新的 [EventListener].
  *
  * e.g.
  * ```kotlin
@@ -36,56 +39,42 @@ import org.slf4j.Logger
  *
  */
 @JvmSynthetic
-public fun EventListener.withMatcher(matcher: suspend EventListenerProcessingContext.() -> Boolean = { true }): MatchableEventListener {
+public fun EventListener.withMatcher(matcher: suspend EventListenerProcessingContext.() -> Boolean = { true }): EventListener {
     if (this is SimpleListener) {
         val old = this.matcher
         return copy(matcher = {
-            old(this) && matcher()
+            old() && matcher()
         })
     }
     
-    if (this is DelegatedMatchableEventListener) {
-        val old = this.matcher
-        return DelegatedMatchableEventListener(this) {
-            old() && matcher()
-        }
-    }
-    
-    return DelegatedMatchableEventListener(this, matcher)
+    return DelegatedEventListener(this, matcher)
 }
 
 
 // region proxy and delegate
 /**
  * 提供一层包装来代理目标监听函数。
- *
- * 如果 [E] 为 [MatchableEventListener], 则返回值也同样为 [MatchableEventListener] 类型。
- * 但是如果不是则返回值类型为普通的 [EventListener] 代理实现。
  */
-public fun <E : EventListener> E.proxy(delegate: suspend EventListenerProcessingContext.(E) -> EventResult): EventListener {
-    // if (this is MatchableEventListener) return MatchableEventListenerProxy(this, delegate)
-    
-    return EventListenerProxy(this, delegate)
-}
-
-// TODO
-/**
- * 提供一层包装来代理目标监听函数。
- */
-public fun <E : MatchableEventListener> E.proxy(
-    handleMatch: suspend EventListenerProcessingContext.(E) -> Boolean,
+public fun <E : EventListener> E.proxy(
+    handleMatch: suspend EventListenerProcessingContext.(E) -> Boolean = { it.match(this) },
     handleInvoke: suspend EventListenerProcessingContext.(E) -> EventResult,
-): MatchableEventListener {
-    return MatchableEventListenerProxy(this, handleMatch, handleInvoke)
+): EventListener {
+    return EventListenerProxy(this, handleMatch, handleInvoke)
 }
 
 
 private class EventListenerProxy<E : EventListener>(
     val listener: E,
-    val handle: suspend EventListenerProcessingContext.(E) -> EventResult,
+    val handleMatch: suspend EventListenerProcessingContext.(E) -> Boolean,
+    val handleInvoke: suspend EventListenerProcessingContext.(E) -> EventResult,
 ) : EventListener by listener {
+    
+    override suspend fun match(context: EventListenerProcessingContext): Boolean {
+        return context.run { handleMatch(this@EventListenerProxy.listener) }
+    }
+    
     override suspend fun invoke(context: EventListenerProcessingContext): EventResult {
-        return context.handle(listener)
+        return context.run { handleInvoke(this@EventListenerProxy.listener) }
     }
     
     
@@ -95,31 +84,10 @@ private class EventListenerProxy<E : EventListener>(
 }
 
 
-private class MatchableEventListenerProxy<E : MatchableEventListener>(
-    val listener: E,
-    val handleMatch: suspend EventListenerProcessingContext.(E) -> Boolean,
-    val handleInvoke: suspend EventListenerProcessingContext.(E) -> EventResult,
-) : MatchableEventListener by listener {
-    
-    override suspend fun match(context: EventListenerProcessingContext): Boolean {
-        return context.run { handleMatch(this@MatchableEventListenerProxy.listener) }
-    }
-    
-    override suspend fun directInvoke(context: EventListenerProcessingContext): EventResult {
-        return context.run { handleInvoke(this@MatchableEventListenerProxy.listener) }
-    }
-    
-    
-    override fun toString(): String {
-        return "ProxiedMatchableEventListener($listener)"
-    }
-}
-
-
-private class DelegatedMatchableEventListener(
+private class DelegatedEventListener(
     val delegate: EventListener,
     val matcher: suspend EventListenerProcessingContext.() -> Boolean,
-) : MatchableEventListener {
+) : EventListener {
     override val id: ID
         get() = delegate.id
     override val logger: Logger
@@ -129,26 +97,12 @@ private class DelegatedMatchableEventListener(
     
     override fun isTarget(eventType: Event.Key<*>): Boolean = delegate.isTarget(eventType)
     
-    private val matcher0: suspend (EventListenerProcessingContext) -> Boolean =
-        if (delegate is MatchableEventListener) {
-            { context -> delegate.match(context) && context.matcher() }
-        } else {
-            { context -> context.matcher() }
-        }
-    
     override suspend fun match(context: EventListenerProcessingContext): Boolean {
-        return matcher0(context)
+        return delegate.match(context) && matcher(context)
     }
     
-    private val invoker0: suspend (EventListenerProcessingContext) -> EventResult =
-        if (delegate is MatchableEventListener) {
-            { context -> delegate.directInvoke(context) }
-        } else {
-            { context -> delegate.invoke(context) }
-        }
-    
-    override suspend fun directInvoke(context: EventListenerProcessingContext): EventResult {
-        return invoker0(context)
+    override suspend fun invoke(context: EventListenerProcessingContext): EventResult {
+        return delegate(context)
     }
     
     override fun toString(): String {
